Consolidate runner capability contracts
