+++ crates/homeboy-lab-runner/src/upgrade_runners/commands.rs
@@ -3,8 +3,8 @@ use crate::Runner;
 use crate::RunnerExecOptions;
 use homeboy_core::Error;
 use homeboy_core::Result;
-use homeboy_lab_runner_contract::RunnerCapabilityPreflight;
-use homeboy_lab_runner_contract::RunnerRequiredTool;
+use homeboy_runner_contract::RunnerCapabilityPreflight;
+use homeboy_runner_contract::RunnerRequiredTool;
 use homeboy_upgrade::upgrade::version_is_newer;
 use homeboy_upgrade::upgrade::InstallMethod;
