+++ crates/homeboy-lab-runner/src/upgrade_runners/source_checkout.rs
@@ -4,11 +4,11 @@ use crate::Runner;
 use crate::RunnerExecOptions;
 use homeboy_core::git::output_allow_empty;
 use homeboy_core::Result;
-use homeboy_lab_runner_contract::RunnerCapabilityPreflight;
 use homeboy_lab_runner_contract::RunnerKind;
-use homeboy_lab_runner_contract::RunnerRequiredTool;
 use homeboy_lab_runner_contract::RunnerWorkspaceSyncMode;
 use homeboy_lab_runner_contract::RunnerWorkspaceSyncOptions;
+use homeboy_runner_contract::RunnerCapabilityPreflight;
+use homeboy_runner_contract::RunnerRequiredTool;
 use homeboy_upgrade::upgrade::current_version;
 use homeboy_upgrade::upgrade::InstallMethod;
 use std::path::Path;
