+++ crates/homeboy-rig/src/capabilities.rs
@@ -7,8 +7,8 @@ use serde::Serialize;
 use super::expand::expand_vars;
 use super::pipeline::{PipelineOutcome, PipelineStepOutcome};
 use super::spec::{ExecutableRequirementSpec, FilesystemAssertionSpec, RigSpec};
-use homeboy_lab_runner_contract::RunnerCapabilityPreflight;
-use homeboy_lab_runner_contract::RunnerToolCapabilityRequirement;
+use homeboy_runner_contract::RunnerCapabilityPreflight;
+use homeboy_runner_contract::RunnerToolCapabilityRequirement;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize)]
 #[allow(
